chore: tests for maximum_borrow_asset_usage_ratio
